[refactor] createcollectionviewcontroller파일 분리
